[REF] web_tour: separation of unit test in multiple files
In this commit, we separate the unit tests that are all grouped in the
tour_service file into files that are more explicit about what is being
tested. We group everything related to the tour automatic feature into
a separate file and we do the same for everything related to tour
interactive.

X-original-commit: 00fb9ba256d8487e038f6d5692cef3e27e01d4a6
